resolves #142 - restructure script command
