+++ src/cli/script/script-describe.js
@@ -1,0 +1,35 @@
+import { Command, Option } from 'commander';
+import * as common from '../cmd_common.js';
+import { getTokens } from '../../ops/AuthenticateOps.js';
+import storage from '../../storage/SessionStorage.js';
+
+const program = new Command('frodo cmd describe');
+
+program
+  .description('Cmd describe.')
+  .helpOption('-h, --help', 'Help')
+  .showHelpAfterError()
+  .addArgument(common.hostArgumentM)
+  .addArgument(common.realmArgument)
+  .addArgument(common.userArgument)
+  .addArgument(common.passwordArgument)
+  .addOption(common.deploymentOption)
+  .addOption(common.insecureOption)
+  .addOption(new Option('-i, --cmd-id <cmd-id>', 'Cmd id.'))
+  .action(
+    // implement command logic inside action handler
+    async (host, realm, user, password, options) => {
+      storage.session.setTenant(host);
+      storage.session.setRealm(realm);
+      storage.session.setUsername(user);
+      storage.session.setPassword(password);
+      storage.session.setDeploymentType(options.type);
+      storage.session.setAllowInsecureConnection(options.insecure);
+      if (await getTokens()) {
+        // code goes here
+      }
+    }
+    // end command logic inside action handler
+  );
+
+program.parse();
